Clean up api.py and added app/utils/api.py
