+++ app/controllers/api.py
@@ -6,28 +6,26 @@ import json
 # Specific to API
 from flask_httpauth import HTTPBasicAuth
 from bson import json_util
-from flask_cors import cross_origin
 
-from app.extensions import cors
-
+# Import the models
 from app.models.users import User
 from app.models.package import Package
 from app.models.token import UserTokens
-# from app import app
+
+from app.utils.api import extract_keys
 
 api = Blueprint('api', __name__)
 
+# This is protect the API routes using HTTP Basic Authentication
 api_auth = HTTPBasicAuth()
 
+# The API route to get a token
 @api.route('/api/user/gettoken', methods=['POST'])
-# @cross_origin(origin='http://localhost:3000')
 def api_gettoken():
     if request.method == 'POST': 
 
-        # if testing using Python Jupyter Notebook
-        # email = request.form.get('email')
-        # password = request.form.get('password')
-
+        # To handle both JSON payloads (commonly used with JavaScript or ReactJS) 
+        # and form data (which can be sent from a Python script or a HTML form).
         try:
             data = request.json
             if data: # if using ReactJS
@@ -37,18 +35,22 @@ def api_gettoken():
             email = request.form.get('email')
             password = request.form.get('password')
 
+        # from OneMap: 400 - You have to enter a valid email address and valid password to generate a token.
         if not email or not password:
-            return jsonify({'error': 'Missing required fields'}), 400
+            return jsonify({'error': 'You have to enter a valid email address and valid password'}), 400 
 
+        # from OneMap: 404 - User is not registered in system.
         user = User.getUser(email=email)
         if not user:
-            return jsonify({'error': 'User not found'}), 404
+            return jsonify({'error': 'User is not registered'}), 404 
 
+        # from OneMap: 401 - Authentication failed, please contact admin at support@onemap.gov.sg
         if not check_password_hash(user.password, password):
-            return jsonify({'error': 'Invalid password'}), 401
+            return jsonify({'error': 'Authentication failed'}), 401 
 
         token =  UserTokens.getToken(email = email)
         
+        # If token exists, return the token
         if token:
             return jsonify({'token': token}), 200
 
@@ -57,7 +59,17 @@ def api_gettoken():
         # Simulated token generation (replace with actual token logic)
         token = generate_password_hash(user.email+datetime_str, method='sha256')
         UserTokens.createToken(email = user.email, token=token)
+        # Return the newly generated token
         return jsonify({'token': token}), 200
+
+# The API route to get all packages  
+@api.route('/api/package/getAllPackages', methods=['POST'])
+@api_auth.login_required
+def getAllPackages():
+    allPackages = Package.getAllPackages()
+    packages_list = [json.loads(json_util.dumps(package.to_mongo())) for package in allPackages]
+    projected_list = [extract_keys(k, idx+1) for idx, k in enumerate(packages_list)]
+    return jsonify({'data': projected_list}), 201
 
 # Protected route for authorized users
 @api.route('/api/protected')
@@ -65,35 +77,7 @@ def api_gettoken():
 def protected():
     return jsonify({'message': 'You are authorized to see this message'}), 201
 
-def extract_keys(dictionary, running_id=1):
-    """
-    Extracts specific keys from a dictionary and adds an "id" key with a running number.
-
-    Args:
-        dictionary: The dictionary to extract keys from.
-        running_id: The initial value for the "id" key (default: 1).
-
-    Returns:
-        A new dictionary containing the extracted keys and the "id" key.
-    """
-    extracted_data = {
-        key: dictionary[key]
-        for key in ["hotel_name", "image_url", "description"]
-        if key in dictionary
-    }
-    extracted_data["id"] = running_id
-    return extracted_data
-    
-@api.route('/api/package/getAllPackages', methods=['POST'])
-# @cross_origin(origin='http://localhost:3000')
-@api_auth.login_required
-def getAllPackages():
-    allPackages = Package.getAllPackages()
-    # products_dict = [json.loads(json_util.dumps(product.to_mongo())) for product in products]
-    packages_list = [json.loads(json_util.dumps(package.to_mongo())) for package in allPackages]
-    projected_list = [extract_keys(k, idx+1) for idx, k in enumerate(packages_list)]
-    return jsonify({'data': projected_list}), 201
-
+# Part of the basic authentication
 @api_auth.verify_password
 def verify_password(email, token):
     user = UserTokens.getToken(email=email)
